Initial commit
This is a new project: HyperList, that is forked from
sergi/virtual-list. It provides bug fixes, improvements, a
standards-compliant ES2015 codebase w/ legacy fallback.

I plan on adding documentation, tests, and maintenance to a stable
launch.

Test now with: `npm install hyperlist` as I've released it on NPM
as a